+++ client/components/Layout.tsx
@@ -1,39 +1,10 @@
 import Head from 'next/head'
-import React, { PropsWithChildren, useCallback, useEffect, useRef } from 'react'
+import React, { PropsWithChildren } from 'react'
 import Header from './Header'
-import { Flex, Box, useColorModeValue, useDisclosure, Text, Slide, Button, Kbd } from "@chakra-ui/react";
-import { Notifications } from './Notifications/Notifications';
-import { NotificationProvider } from '../providers/NotificationProvider';
-import { ArrowForwardIcon } from '@chakra-ui/icons';
+import { Flex, Box, useColorModeValue } from "@chakra-ui/react";
+import { NotificationSidebar } from './Notifications/NotificationSidebar';
 
 export default function Layout({ children }: PropsWithChildren<{}>) {
-    // TODO: Move into provider / component
-    const { isOpen: showNotifications, onToggle: toggleNotifications, onClose } = useDisclosure();
-
-    const notificationBarRef = useRef<HTMLDivElement | null>(null);
-    // Close on outside click
-    const handleClick = useCallback((event) => {
-        if (notificationBarRef.current && !notificationBarRef.current.contains(event.target)) {
-            onClose();
-        }
-    }, [onClose]);
-
-    // Close on escape
-    const handleKeyDown = useCallback((event) => {
-        if (event.keyCode === 27) {
-            onClose();
-        }
-    }, [onClose]);
-
-    useEffect(() => {
-        document.addEventListener('click', handleClick, true);
-        document.addEventListener('keydown', handleKeyDown, true);
-        return () => {
-            document.removeEventListener('click', handleClick, true);
-            document.removeEventListener('keydown', handleKeyDown, true);
-        };
-    }, [handleClick, handleKeyDown]);
-
     return (
         <Flex
             direction="column"
@@ -46,43 +17,8 @@ export default function Layout({ children }: PropsWithChildren<{}>) {
                 <meta name="description" content="Overview of game release dates, prices and news" />
                 <link rel="icon" href="/favicon.ico" />
             </Head>
-            <NotificationProvider>
-                <Header toggleNotifications={toggleNotifications} />
-                <Slide
-                    direction="right"
-                    in={showNotifications}
-                    style={{
-                        marginTop: "4rem",
-                        zIndex: 4
-                    }}
-                >
-                    <Box
-                        ref={notificationBarRef}
-                        position="absolute"
-                        zIndex="4"
-                        bg="gray.800"
-                        right="0"
-                        height="calc(100vh - 4rem)"
-                        width="25rem"
-                    >
-                        <Flex p="0.5rem" align="center" justifyContent="space-between">
-                            <Text fontWeight="bold" ml="0.5rem">
-                                Notifications
-                            </Text>
-                            <Button
-                                rightIcon={<ArrowForwardIcon />}
-                                onClick={onClose}
-                                size="sm"
-                                colorScheme='teal'
-                                variant='ghost'
-                            >
-                                Collapse<Kbd ml="0.5rem">Esc</Kbd>
-                            </Button>
-                        </Flex>
-                        <Notifications />
-                    </Box>
-                </Slide>
-            </NotificationProvider>
+            <Header />
+            <NotificationSidebar />
             <Box
                 mt="4rem"
                 pt="2rem"
