+++ client/components/Notifications/Notification.tsx
@@ -1,0 +1,96 @@
+import { Box, Flex, Text } from "@chakra-ui/layout";
+import dayjs from "dayjs";
+import { NotificationDto, NotificationType } from "@game-watch/shared";
+import { SourceTypeLogoSmall } from "../InfoSource/SourceTypeLogo";
+import { IconButton, Tooltip } from "@chakra-ui/react";
+import { CheckCircleIcon } from "@chakra-ui/icons";
+import React, { useMemo } from "react";
+
+const NotificationTypeNames: Record<NotificationType, string> = {
+    [NotificationType.NewStoreEntry]: "New store entry",
+    [NotificationType.ReleaseDateChanged]: "New release date",
+    [NotificationType.GameReleased]: "Game released",
+    [NotificationType.GameReduced]: "Game reduced",
+    [NotificationType.NewMetacriticRating]: "Game rated",
+}
+
+function isNotificationOfType<T extends NotificationType, TValue extends NotificationType>(
+    notification: NotificationDto<NotificationType>,
+    type: TValue
+): notification is NotificationDto<Extract<T, TValue>> {
+    return notification.type === type;
+}
+
+const getNotificationText = (notification: NotificationDto) => {
+    const { game } = notification;
+
+    // TODO: Scroll-To-Game Effect with simple html links?
+    const gameName = <Text fontWeight="bold" display="inline">{game.name || game.search}</Text>;
+
+    if (isNotificationOfType(notification, NotificationType.GameReduced)) {
+        const { final, initial } = notification.data;
+        return <>{gameName} {`was reduced from ${initial}€ to ${final}€!`}</>;
+    }
+
+    if (isNotificationOfType(notification, NotificationType.ReleaseDateChanged)) {
+        const formattedDate = dayjs(notification.data.releaseDate).format("DD. MMM. YYYY")
+        return <>{gameName} {`will be released on (${formattedDate})!`}</>;
+    }
+
+    switch (notification.type) {
+        case NotificationType.NewStoreEntry:
+        case NotificationType.GameReleased:
+            return <>{gameName} is now available!</>;
+        case NotificationType.NewMetacriticRating:
+            return <>{gameName} received a rating!</>;
+        default:
+            return <>Unknown Notification</>;
+    }
+}
+
+const NotificationComponent: React.FC<{
+    notification: NotificationDto,
+    markNotificationAsRead: (id: string) => Promise<void>
+}> = ({ notification, markNotificationAsRead }) => {
+
+    const notificationText = useMemo(() => getNotificationText(notification), [notification]);
+
+    return (
+        <Flex
+            direction="column"
+            mb="0.5rem"
+            mx="0.5rem"
+            p="1rem"
+            rounded="lg"
+            bg={'gray.700'}
+            position="relative"
+        >
+            <Flex align="center" pb="1rem">
+                <Tooltip label={dayjs(notification.createdAt).format("DD.MM. HH:mm")} placement="right">
+                    <Text fontSize="lg" fontWeight="bold">
+                        {NotificationTypeNames[notification.type]}
+                    </Text>
+                </Tooltip>
+                <Box ml="1rem">
+                    <a href={notification.infoSource.data?.url} target="_blank" rel="noreferrer">
+                        {SourceTypeLogoSmall[notification.infoSource.type]}
+                    </a>
+                </Box>
+            </Flex>
+            <Box position="absolute" right="0.5rem" top="0.5rem">
+                <IconButton
+                    icon={<CheckCircleIcon />}
+                    onClick={() => markNotificationAsRead(notification.id)}
+                    size="sm"
+                    variant="ghost"
+                    aria-label='Delete'
+                />
+            </Box>
+            <Box>
+                {notificationText}
+            </Box>
+        </Flex>
+    )
+}
+
+export const Notification = React.memo(NotificationComponent);
