+++ client/pages/_app.tsx
@@ -3,14 +3,17 @@ import type { AppProps } from 'next/app'
 import { ChakraProvider } from "@chakra-ui/react"
 import Layout from "../components/Layout";
 import theme from '../theme';
+import { NotificationProvider } from '../providers/NotificationProvider';
 
 function MyApp({ Component, pageProps }: AppProps) {
-
   return (
     <ChakraProvider theme={theme}>
-      <Layout>
-        <Component {...pageProps} />
-      </Layout>
+      <NotificationProvider>
+        <Layout>
+          <Component {...pageProps} />
+        </Layout>
+      </NotificationProvider>
+
     </ChakraProvider>
   );
 }
